requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.15
httpx>=0.24.1
feedparser>=6.0.10
pyyaml>=6.0
python-dotenv>=1.0.0
